Updates Pokemon.py
I'd been using an inefficient way of submitting input to a form

In this doc, I've added ways to use the form in order to get data from the user in an efficient manner. 
The names I get from of the user are stored in the form and passed to a variable called pokemonName 
If the name entered is not in pokemonLIst List then it gets appended to it,  and a flash message is sent. 
Otherwise, the name isn't appended to the list and it triggers a failure flash message. 

The rendered page is then rendered with an accessible variable called form that represents the form of the 
pkmnForm's classes and variables.